% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesint_table.R
\name{bayesint_table}
\alias{bayesint_table}
\title{Create a table of maximum likelihood and Bayesian credible interval positions for a QTL object.}
\usage{
bayesint_table(qtl_object, model_fit)
}
\arguments{
\item{qtl_object}{A QTL object from r/qtl.}

\item{model_fit}{A fitQTL object from r/qtl.}
}
\value{
A data frame listing QTL name, the maximum likelihood position of the QTL and the upper and lower
confidence intervals.
}
\description{
Create a table of maximum likelihood and Bayesian credible interval positions for a QTL object.
}
