% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_qtl.R
\name{cluster_qtl}
\alias{cluster_qtl}
\title{Group QTL from different experiments into blocks.}
\usage{
cluster_qtl(chr, qtl_list, model_fit_list, threshold = NULL)
}
\arguments{
\item{chr}{A numeric vector listing which chromosomes are to be plotted.}

\item{qtl_list}{A list of QTL objects from r/qtl (i.e list(object1, object2, ...))}

\item{threshold}{Maximum length in centimorgans above which QTL are considered to be
poorly defined and excluded from clustering.}
}
\description{
Loci in a list of QTL objects are grouped by position.
Loci are declared to be colocalised if the confidence intervals on locus position overlap.
When the position of a QTL is poorly defined and confidence intervals are large, the locus
colocalises with many other loci, probably erroneously. To combat this, set a maximum
length using the argument threshold; QTL with confidence intervals longer than this will
be excluded from clustering.
}

