% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_qtl_summary.R
\name{cluster_qtl_summary}
\alias{cluster_qtl_summary}
\title{Make a summary table of clusters of QTL.}
\usage{
cluster_qtl_summary(qtl_clusters)
}
\arguments{
\item{qtl_clusters}{Data.frame containing information on clustered QTL,
usually from the output of cluster_qtl.}
}
\value{
Data.frame listing information for each cluster. Information
includes the chromosome number, number of QTL in the cluster, then the range
and weighted mean values for position (cM), effect size and precentage of the
variation explained.
}
\description{
Creates a data.frame summarising information about groups of QTL clustered
using cluster_qtl(). Clusters will be labelled with the chromosome and index,
with an optional prefix.
}
