% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_box.R
\name{make_box}
\alias{make_box}
\title{Convert the output of cluster_qtl_summary to a list of box coordinates}
\usage{
make_box(dat)
}
\arguments{
\item{dat}{Output of cluster_qtl_summary.}
}
\value{
Dataframe listing chromosome name, and upper and lower bounds of the
box.
}
\description{
Convert the output of cluster_qtl_summary to a list of box coordinates
}
