% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTL.R
\name{plotQTL}
\alias{plotQTL}
\title{Create an object needed to create QTL plots.}
\usage{
plotQTL(chr = 1:5, marker_locations, nlanes = 2, left_gap = 2, right_gap = 2)
}
\arguments{
\item{chr}{a numeric vector listing which chromosomes are to be plotted.}

\item{marker_locations}{A data.frame listing marker positions on each chromosome.
Should include a column labelled 'Position' and another labelled 'Chromosome'.}

\item{nlanes}{the number of lanes to plot for each chromosome.}

\item{left_gap, right_gap}{Additional space on the left- and right-hand sides
of each chromosome for additional labels.}
}
\value{
A list of two tables detailing:
\enumerate{
\item  Chromosomes to be plotted
\item Number of chromosome tracks to be included
\item Number of lanes on each track
\item Lists of marker positions for each track
\item Lengths of each track
\item Gap to the left of each track
\item Gap to the right of each track
\item Table of x-axis poisitions of the centres of each lane.
\item Table of x-axis poisitions of the margins of each lane.
\item Optional list of labels for each lane.
}
}
\description{
Creates a list of information needed to plot QTL across a genome, for all or
only a subset of chromsomes.
}
\details{
For each chromosome specified a 'track' is created, which is further divided
into lanes for plotting QTL from different years or experiments. QTL
positions and their confidence intervals are plotted to the centre of these
lanes, and dividing lanes plotted between each lane. These lane centres and
margin positions can be used to add extra information, such as lane labels
above the chromosome.

The gap between chromosomes can be adjusted to include further embellishments,
and the plot  will rescale automatically.
}
