% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTL_base.R
\name{plotQTL_base}
\alias{plotQTL_base}
\title{Set up an empty plot for QTL positions}
\usage{
plotQTL_base(plotQTL, by = 20, maxy = 11, miny = 11, track_labels = NULL, ...)
}
\arguments{
\item{plotQTL}{A plotQTL object from plotQTL().}

\item{by}{Intervals between y-axis tick labels. Defaults to 20.}

\item{maxy, miny}{Upper and lower limits of the y-axis.}

\item{track_labels}{Vector of labels to adorn the bottom of each track.}

\item{...}{Additional arguments passed to graphical parameters.}
}
\value{
An empty plot to which QTL tracks can be added.
}
\description{
Creates an empty genome mapping for plotting positions and confidence intervals
of QTL inferred using r/qtl
}
