% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTL_box.R
\name{plotQTL_box}
\alias{plotQTL_box}
\title{Add box to a plotQTL plot.}
\usage{
plotQTL_box(plotQTL, track, upper, lower, left_lane = NULL,
  right_lane = NULL, lwd = 0.5, border = "gray50", col = "gray76", ...)
}
\arguments{
\item{plotQTL}{A plotQTL object from plotQTL().}

\item{track}{Integer indicating which chromosome track in the plotQTL object to plot to.}

\item{upper}{Marker position for the top of the box.}

\item{lower}{Marker position for the top of the box.}

\item{left_lane}{Integer indicating the left-hand lane margin of the box.}

\item{right_lane}{Integer indicating the right-hand lane margin of the box.}

\item{lwd}{Width of the border.}

\item{border}{Colour of the border.}

\item{col}{Colour to fill the box.}
}
\description{
Add a rectangle to a plotQTL plot, onto which QTL positions can be plotted.
For example, boxes could denote regions in which QTL were found in previous studies.
By default boxes span all lanes of the chromosome. Boxes can be restricted to particular
lanes by specifying the lane margins in left_lane and right_lane.
}

