% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTL_boxes.R
\name{plotQTL_boxes}
\alias{plotQTL_boxes}
\title{Add box to a plotQTL plot.}
\usage{
plotQTL_boxes(plotQTL, boxes, margins = "full", ...)
}
\arguments{
\item{plotQTL}{A plotQTL object from plotQTL().}

\item{boxes}{A dataframe listing chomosome, lower bound and upper bounds for 
the box. This can be generated using cluster_qtl().}

\item{margins}{Optional vector of two elements giving the left- and right-most
margins for the box. If not used, the whole width of the chromosome will be
used.}

\item{...}{Additional parameters passed to graphical functions.}
}
\description{
Add a rectangle to a plotQTL plot, onto which QTL positions can be plotted.
For example, boxes could denote regions in which QTL were found in previous studies.
By default boxes span all lanes of the chromosome. Boxes can be restricted to particular
lanes by specifying the lane margins in left_lane and right_lane.
}
