% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTL_ladder.R
\name{plotQTL_ladder}
\alias{plotQTL_ladder}
\title{Plot marker positions in a plotQTL plot.}
\usage{
plotQTL_ladder(
  plotQTL,
  marker_tick_width = 0.5,
  col = "gray50",
  lty = "dashed",
  ...
)
}
\arguments{
\item{plotQTL}{A plotQTL object from plotQTL().}

\item{marker_tick_width}{Width of marker labels.}

\item{col}{Colour for lane margins.}

\item{lty}{Line style.}

\item{...}{Further graphical parameters passed to \code{segments}.}
}
\description{
Plots a solid axis with marker locations to the left of each chromosome track,
and dashed grey dividers between lanes.
}
