% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTL_lanes.R
\name{plotQTL_lanes}
\alias{plotQTL_lanes}
\title{Plot lane margins in a plotQTL plot.}
\usage{
plotQTL_lanes(plotQTL, marker_tick_width = 0.5, col = "gray50",
  lty = "dashed")
}
\arguments{
\item{plotQTL}{A plotQTL object from plotQTL().}

\item{marker_tick_width}{Width of marker labels.}

\item{col}{Colour for lane margins}
}
\description{
Plots a solid axis with marker locations to the left of each chromosome track,
and dashed grey dividers between lanes.
}

