% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQTL_qtl.R
\name{plotQTL_qtl}
\alias{plotQTL_qtl}
\title{Plot QTL positions.}
\usage{
plotQTL_qtl(plotQTL, qtl_object, model_fit, lane, fill_cutoff = 0, col = 1,
  ...)
}
\arguments{
\item{plotQTL}{A plotQTL object from plotQTL().}

\item{model_fit}{A fitQTL object from r/qtl.}

\item{lane}{Integer indicating which lane to plot to.}

\item{fill_cutoff}{Threshold length of a confidence interval on QTL positions, above which
points will be empty, and below which points will be filled.}

\item{col}{Plotting colour.}

\item{...}{Additional plotting commands.}

\item{qtlobject}{A QTL object from r/qtl.}
}
\description{
Plot QTL positions and confidence intervals from an r/qtl object onto a plotQTL plot.
Maximum-likelihood positions are labelled with an upward arrow if the reference allele
increases the phenotype, and vice versa.
}
\details{
QTL with narrow confidence intervals are plotted with filled arrows. If confidence
intervals are greater than a threshold number of centimorgans, arrows are plotted empty.
This can be disabled by setting fill_cutoff to a larger number than the longest interval.
}

