% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_dataframe.R
\name{qtl_dataframe}
\alias{qtl_dataframe}
\title{Concatenate data tables on genotypes and phenotypes into a format compatible with R/qtl.}
\usage{
qtl_dataframe(phenotypes, genotypes, phenotype_id_column = 1,
  genotype_id_column = 1)
}
\arguments{
\item{phenotypes}{A table of phenotype data.}

\item{genotypes}{A table of genotype data, including marker information in the first two rows.}

\item{phenotype_id_column}{The column in \code{phenotype} indicating line names.}

\item{genotype_id_column}{The column in \code{genotype} indicating line names to match up with
those for the phenotype table.}
}
\value{
# a dataframe of the format for importing into R/qtl.
}
\description{
Match datatables for phenotypes and genotypes for a set of lines, and combine these into a 
single dataframe that r/qtl can read.
}
