% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtleffects.R
\name{qtleffects}
\alias{qtleffects}
\title{Create qtleffects object}
\usage{
qtleffects(chr, marker_locations)
}
\arguments{
\item{chr}{a numeric vector listing which chromosomes are to be plotted.}

\item{marker_locations}{A data.frame listing marker positions on each chromosome.
Should include a column labelled 'Position' and another labelled 'Chromosome'.}

\item{nlanes}{the number of lanes to plot for each chromosome. Defaults to two for Sweden vs. Italy.
left_gap, right_gap left- and right-hand margin around each chromosome.}
}
\value{
A qtleffects object listing chromosomes to be plotted, x-values for marker positions,
and x-values for the start of each chromosome.
}
\description{
Creates an object needed to plot effect sizes of QTL from r/qtl.
}

