% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtleffects_base.R
\name{qtleffects_base}
\alias{qtleffects_base}
\title{Base plot for qtleffects}
\usage{
qtleffects_base(qtleffects, y, ylabel = NULL, tick.length = NULL,
  tick.lwd = 0.5, box = TRUE)
}
\arguments{
\item{qtleffects}{A qtleffects object from qtleffects() denoting x-value positions.}

\item{y}{A vector of two elements denoting minimum and maximum y-vales. Equivalent to ylim argument.}

\item{ylabel}{Axis label for the y-axis. If NULL, this deaults to 'Effect size'.}

\item{tick.length}{Length of ticks for marker locations. If NULL, defaults to 3% of the total.
y-axis length. Negative values will anchor ticks above the minimum y value.}

\item{tick.lwd}{Line widths for marker locations.}

\item{box}{If TRUE, a border will be drawn around the plot.}
}
\description{
Creates an empty plot for plotting QTL effect sizes from a qtleffects object.
}
