% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underlined.R
\name{underlined}
\alias{underlined}
\title{Underlined text}
\usage{
underlined(x, y, label, cex = 1, ...)
}
\arguments{
\item{x}{X-axis position for text.}

\item{y}{Y-axis position for text.}

\item{label}{Text to be plotted.}

\item{cex}{A numerical value giving the amount by which plotting text and
symbols should be magnified relative to the default. See `?par`.}

\item{...}{Further arguments to be passed to text().}
}
\description{
Add underlined text to a plot.
}
\details{
This is a workaround to get unerlined text on a plot. This is edited from a
post by Uwe Ligges at
\url{https://tolstoy.newcastle.edu.au/R/help/02a/0471.html}.

It will not work if the text is at an angle (i.e. \code{srt != 0}), or if
the axis is on a log scale.
}
