% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_qtl.R
\name{cluster_qtl}
\alias{cluster_qtl}
\title{Group QTL from different experiments into blocks.}
\usage{
cluster_qtl(chr, qtl_list, model_fit_list, threshold = NULL,
  hard_boundary = TRUE, qtl_labels = "QTL")
}
\arguments{
\item{chr}{A numeric vector listing which chromosomes are to be plotted.}

\item{qtl_list}{A list of QTL objects from r/qtl (i.e list(object1, object2, ...)).}

\item{model_fit_list}{A list of model fit objects (i.e. objects outputted from `fit.qtl`).}

\item{threshold}{Maximum length in centimorgans above which QTL are considered to be
poorly defined and excluded from clustering.}

\item{hard_boundary}{Logical. Determines whether to separate pairs QTL that
overlap at only a single marker, i.e. the end of the credible interval of the
first QTL falls at the same marker where the credible interval begins. If
TRUE, these QTL are treated as distinct. Defaults to TRUE.}

\item{qtl_labels}{Prefix for labels for QTL. Usually a string. Defaults to "QTL".}
}
\value{
Two data frames:
\enumerate{
\item{1. A summary table of each cluster, indicating the
mean and the maximum extent of 95% credible intervals, effect sizes, and
percentage of total phenotypic variance explained for all QTL in the cluster.
See \code{?cluster_qtl_summary}.}
\item{2. Full details of all the QTL, indicating which clutser they belong to. }
}
}
\description{
Loci in a list of QTL objects are grouped by position.
Loci are declared to be colocalised if the confidence intervals on locus
position overlap. When the position of a QTL is poorly defined and confidence
intervals are large, the locus colocalises with many other loci, probably
erroneously. To combat this, set a maximum length using the argument
threshold; QTL with confidence intervals longer than this will be excluded
from clustering.
}
