% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtleffects_points.R
\name{qtleffects_points}
\alias{qtleffects_points}
\title{Points for QTL effect sizes.}
\usage{
qtleffects_points(qtleffects, qtl_object, model_fit, type = "effect", ...)
}
\arguments{
\item{qtleffects}{A qtleffects object from qtleffects() denoting x-value positions.}

\item{model_fit}{A fitQTL object from r/qtl.}

\item{type}{Variable to be plotted. This should either be 'effect' to plot raw effect
sizes, or 'PVE' to plot percentages of variance explained.}

\item{...}{further arguments passed to points.}

\item{qtlobject}{A QTL object from r/qtl.}
}
\description{
Add points for QTL effect sizes to an existing qtleffects plot.
These can be raw effect sizes or percent variance explained (PVE).
For the former, confidence intervals are also plotted.
}

