% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underlined.R
\name{underlined}
\alias{underlined}
\title{Add underlined text to a plot.}
\usage{
underlined(x, y, label, cex = 1, ...)
}
\arguments{
\item{x}{X-axis position for text.}

\item{y}{Y-axis position for text.}

\item{label}{Text to be plotted.}

\item{cex}{A numerical value giving the amount by which plotting text and
symbols should be magnified relative to the default. See `?par`.}

\item{...}{Further arguments to be passed to text().}
}
\description{
Add underlined text to a plot.
}
